This commit introduces jieba.initialize(sqlite = True|"/path/to/jieba.db")
Sqlite is implemented using SqlAlchemy to create two, indexed sqlite tables
in a single db file (default /tmp/jieba.db).

If the user does not choose sqlite, nothing will happen. Just default to
the in-memory version of Jieba.

Performance seems very good and Memory usage is *much* better.

I am using this in a mobile application. (The final DB file is about 24 MB,
which can be downloaded to the mobile application using after it is installed.)

Comments welcome!