Add only logging.NullHandler to default_logger
This is a library, not an application, and it should be up to the
end user, not the developer, what the log level and handler is.

Setting up the logger with a DEBUG level and sys.stderr handler
is not a widely used practice in Python libraries.  The common
practice is to add only NullHandler:

> An instance of this handler could be added to the top-level
> logger of the logging namespace used by the library
> (if you want to prevent your library’s logged events
> being output to sys.stderr in the absence of logging configuration).

See also:
https://docs.python-guide.org/writing/logging/#logging-in-a-library